Changes - New interface `BaseCommandExecuteData` - Rename `CommandHaltReasonBase` -> `BaseCommandHaltData`
