+++ src/reciple/types/commands.ts
@@ -1,8 +1,9 @@
 import { MessageCommandExecuteData, MessageCommandHaltData } from '../classes/builders/MessageCommandBuilder';
 import { SlashCommandExecuteData, SlashCommandHaltData } from '../classes/builders/SlashCommandBuilder';
 import { MessageCommandOptionManager } from '../classes/MessageCommandOptionManager';
-import { CommandBuilderType, AnyCommandExecuteData } from '../types/builders';
 import { CooledDownUser } from '../classes/CommandCooldownManager';
+import { RecipleClient } from '../classes/RecipleClient';
+import { CommandBuilderType } from '../types/builders';
 
 /**
  * Any command halt data
@@ -15,9 +16,21 @@ export type AnyCommandHaltData = SlashCommandHaltData|MessageCommandHaltData;
 export type CommandHaltData<T extends CommandBuilderType> = CommandErrorData<T>|CommandCooldownData<T>|(T extends CommandBuilderType.SlashCommand ? never : CommandInvalidArguments<T>|CommandMissingArguments<T>)|CommandMissingMemberPermissions<T>|CommandMissingBotPermissions<T>;
 
 /**
+ * Any command execute data
+ */
+export type AnyCommandExecuteData = SlashCommandExecuteData|MessageCommandExecuteData;
+
+/**
+ * Command execute data
+ */
+export interface BaseCommandExecuteData {
+    client: RecipleClient<true>;
+}
+
+/**
  * Command halt reason base
  */
-export interface CommandHaltReasonBase<T extends CommandBuilderType> {
+export interface BaseCommandHaltData<T extends CommandBuilderType> {
     executeData: T extends CommandBuilderType.SlashCommand
                     ? SlashCommandExecuteData
                     : T extends CommandBuilderType.MessageCommand 
@@ -25,25 +38,25 @@ export interface CommandHaltReasonBase<T extends CommandBuilderType> {
                         : AnyCommandExecuteData
 }
 
-export interface CommandErrorData<T extends CommandBuilderType> extends CommandHaltReasonBase<T> {
+export interface CommandErrorData<T extends CommandBuilderType> extends BaseCommandHaltData<T> {
     reason: CommandHaltReason.Error;
     error: any;
 }
-export interface CommandCooldownData<T extends CommandBuilderType> extends CommandHaltReasonBase<T>,CooledDownUser {
+export interface CommandCooldownData<T extends CommandBuilderType> extends BaseCommandHaltData<T>,CooledDownUser {
     reason: CommandHaltReason.Cooldown;
 }
-export interface CommandInvalidArguments<T extends CommandBuilderType> extends CommandHaltReasonBase<T> {
+export interface CommandInvalidArguments<T extends CommandBuilderType> extends BaseCommandHaltData<T> {
     reason: CommandHaltReason.InvalidArguments;
     invalidArguments: MessageCommandOptionManager;
 }
-export interface CommandMissingArguments<T extends CommandBuilderType> extends CommandHaltReasonBase<T> {
+export interface CommandMissingArguments<T extends CommandBuilderType> extends BaseCommandHaltData<T> {
     reason: CommandHaltReason.MissingArguments;
     missingArguments: MessageCommandOptionManager;
 }
-export interface CommandMissingMemberPermissions<T extends CommandBuilderType> extends CommandHaltReasonBase<T> {
+export interface CommandMissingMemberPermissions<T extends CommandBuilderType> extends BaseCommandHaltData<T> {
     reason: CommandHaltReason.MissingMemberPermissions;
 }
-export interface CommandMissingBotPermissions<T extends CommandBuilderType> extends CommandHaltReasonBase<T> {
+export interface CommandMissingBotPermissions<T extends CommandBuilderType> extends BaseCommandHaltData<T> {
     reason: CommandHaltReason.MissingBotPermissions;
 }
